Refactor config validator to use ts-to-zod
