+++ README.md
@@ -212,7 +212,9 @@ $ yarn test --watch
 And a playground inside `example`, buildable with the following command:
 
 ```sh
-$ yarn example
+$ yarn gen:example
 ```
+
+Last note, if you are updating `src/config.ts`, you need to run `yarn gen:config` to have generate the schemas of the config (`src/config.zod.ts`) (Yes, we are using the tool to build itself #inception)
 
 Have fun!
