+++ package.json
@@ -22,8 +22,9 @@
     "prepublishOnly": "yarn test:ci && rimraf lib && yarn build",
     "format": "eslint **/*.{js,jsx,ts,tsx} --fix && prettier **/*.{js,jsx,ts,tsx,json} --write",
     "test": "jest",
-    "test:ci": "jest --ci --coverage && yarn example && tsc --noEmit",
-    "example": "./bin/run example/heros.ts example/heros.zod.ts",
+    "test:ci": "jest --ci --coverage && yarn gen:example && yarn gen:config && tsc --noEmit",
+    "gen:example": "./bin/run --config example",
+    "gen:config": "./bin/run --config config",
     "prepare": "husky install"
   },
   "files": [
