+++ src/config.zod.ts
@@ -1,0 +1,32 @@
+// Generated by ts-to-zod
+import { z } from "zod";
+
+export const getSchemaNameSchema = z
+  .function()
+  .args(z.string())
+  .returns(z.string());
+
+export const nameFilterSchema = z
+  .function()
+  .args(z.string())
+  .returns(z.boolean());
+
+export const configSchema = z.object({
+  input: z.string().optional(),
+  output: z.string().optional(),
+  skipValidation: z.boolean().optional(),
+  maxRun: z.number().optional(),
+  nameFilter: nameFilterSchema.optional(),
+  getSchemaName: getSchemaNameSchema.optional(),
+  keepComments: z.boolean().optional().default(false),
+});
+
+export const configsSchema = z.array(
+  configSchema.and(
+    z.object({
+      name: z.string(),
+    })
+  )
+);
+
+export const tsToZodConfigSchema = z.union([configSchema, configsSchema]);
