+++ ts-to-zod.config.js
@@ -1,0 +1,13 @@
+/**
+ * ts-to-zod configuration.
+ *
+ * @type {import("./src/config").TsToZodConfig}
+ */
+module.exports = [
+  {
+    name: "example",
+    input: "example/heros.ts",
+    output: "example/heros.zod.ts",
+  },
+  { name: "config", input: "src/config.ts", output: "src/config.zod.ts" },
+];
